+++ web/src/components/ui/NpcStationsPane.tsx
@@ -84,6 +84,8 @@ export function NpcStationsPane({ eveSystemId }: { eveSystemId: number | null })
 
   useEffect(() => {
     if (!eveSystemId) return;
+    // Deliberate: clears this pane's own state when the record it shows changes.
+    // eslint-disable-next-line react-hooks/set-state-in-effect
     setStations([]);
     setLoading(true);
     fetchStations(eveSystemId)
