+++ web/src/components/ui/ScoutConnectionsPane.tsx
@@ -6,7 +6,8 @@ import { useWormholeTypes } from '../../hooks/useWormholeTypes';
 import { whSizeForType, whSizeShort } from '../../utils/wormholeSize';
 import { useRouteOrigin } from '../../hooks/useRouteOrigin';
 import { useRoute } from '../../hooks/useRoute';
-import { setWaypoint, RouteSquares, canSetAutopilot } from './routeUi';
+import { RouteSquares } from './routeUi';
+import { setWaypoint, canSetAutopilot } from '../../utils/routeActions';
 import { useSystemAlias } from '../../hooks/useSystemAlias';
 import { truesecColor } from '../../utils/truesec';
 import { useMapStore } from '../../store/mapStore';
